@@ -3,7 +3,7 @@ package com.example.cache
 import android.media.AudioManager
 import android.media.MediaPlayer
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_main.*
 import org.jetbrains.anko.attempt
 import org.jetbrains.anko.doAsync
